pci: extend MSI interrupt vector number to 32 bits
Also increase MSI message data size to 32 bits according to PCIe spec.

- MSI-X supports up to 2048 interrupts that do not fit to `uint8`.

- Non-x86 systems may use separate interrupt vector ranges for
hard-wired interrupts and MSI interrupts so `uint8` is not enough to
represent all of them.

+++ src/add-ons/kernel/busses/i2c/pch/pch_i2c.h
@@ -53,7 +53,7 @@ acpi_status pch_i2c_scan_bus_callback(acpi_handle object, uint32 nestingLevel,
 
 struct pch_i2c_crs {
 	uint16	i2c_addr;
-	uint8	irq;
+	uint32	irq;
     uint8	irq_triggering;
 	uint8	irq_polarity;
 	uint8	irq_shareable;
